+++ crates/miden-objects/src/account/delta/storage.rs
@@ -9,7 +9,7 @@ use super::{
     Word,
 };
 use crate::{
-    Digest, EMPTY_WORD,
+    Digest, EMPTY_WORD, Felt, ZERO,
     account::{AccountStorage, StorageMap, StorageSlot},
 };
 // ACCOUNT STORAGE DELTA
@@ -107,6 +107,52 @@ impl AccountStorageDelta {
     /// Returns an iterator of all the updated storage slots.
     fn updated_slots(&self) -> impl Iterator<Item = (&u8, &Word)> + '_ {
         self.values.iter().filter(|&(_, value)| value != &EMPTY_WORD)
+    }
+
+    /// Appends the storage slots delta to the given `elements` from which the delta commitment will
+    /// be computed.
+    ///
+    /// TODO: Make num_slots part of this struct.
+    pub(super) fn append_delta_elements(&self, elements: &mut Vec<Felt>, num_slots: u8) {
+        const DOMAIN_VALUE: Felt = Felt::new(2);
+        const DOMAIN_MAP: Felt = Felt::new(3);
+
+        for slot_idx in 0..num_slots {
+            let slot_idx_felt = Felt::from(slot_idx);
+
+            // The storage delta ensures that the value slots and map slots do not have overlapping
+            // slot indices, so at most one of them will return `Some` for a given slot index.
+            match self.values.get(&slot_idx) {
+                Some(new_value) => {
+                    elements.extend_from_slice(&[DOMAIN_VALUE, slot_idx_felt, ZERO, ZERO]);
+                    elements.extend_from_slice(new_value);
+                },
+                None => {
+                    if let Some(map_delta) = self.maps().get(&slot_idx) {
+                        if map_delta.is_empty() {
+                            continue;
+                        }
+
+                        for (key, value) in map_delta.entries() {
+                            elements.extend_from_slice(key.as_elements());
+                            elements.extend_from_slice(value);
+                        }
+
+                        let num_changed_entries = Felt::try_from(map_delta.num_entries()).expect(
+                            "number of changed entries should not exceed max representable felt",
+                        );
+
+                        elements.extend_from_slice(&[
+                            DOMAIN_MAP,
+                            slot_idx_felt,
+                            num_changed_entries,
+                            ZERO,
+                        ]);
+                        elements.extend_from_slice(&EMPTY_WORD);
+                    }
+                },
+            }
+        }
     }
 }
 
@@ -224,8 +270,13 @@ impl StorageMapDelta {
         Self(map)
     }
 
-    /// Returns a reference to the updated leaves in this storage map delta.
-    pub fn leaves(&self) -> &BTreeMap<Digest, Word> {
+    /// Returns the number of changed entries in this map delta.
+    pub fn num_entries(&self) -> usize {
+        self.0.len()
+    }
+
+    /// Returns a reference to the updated entries in this storage map delta.
+    pub fn entries(&self) -> &BTreeMap<Digest, Word> {
         &self.0
     }